+++ CHANGELOG.md
@@ -7,8 +7,13 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+
+- The ability to delete your account.
+
 ### Changed
 
+- Move the passwords into the accounts.
 - hnefatafl-client: make the game display prettier.
 - Make the board display as capital letters.